requirements.txt
google-generativeai>=0.5.0
sentence-transformers>=2.7.0
numpy>=1.26.0
python-dotenv>=1.0.0

# REST API
fastapi>=0.111.0
uvicorn>=0.29.0
python-multipart>=0.0.9

# Tests
pytest>=8.0.0
